+++ test/kodo_tests.cpp
@@ -8,10 +8,15 @@
 
 #include <gtest/gtest.h>
 
+#include "test_profile.hpp"
+
 GTEST_API_ int main(int argc, char **argv)
 {
     srand(static_cast<uint32_t>(time(0)));
 
     testing::InitGoogleTest(&argc, argv);
+
+    parse_test_arguments(argc, argv);
+
     return RUN_ALL_TESTS();
 }