+++ NEWS.rst
@@ -11,6 +11,7 @@ Latest
   on embedded devices
 * Minor: Added wrapper for ``src/copy_payload_decoder.hpp``. The wrapper
   helps fiting the copy payload layer in the layer stack in an easy way.
+* Patch: Disabled the makefile example on Windows
 
 18.0.0
 ------